requirements.txt
tensorflow==2.13.0
streamlit==1.28.0
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0